third_party/blink/web_tests/external/wpt/css/selectors/invalidation/part-dir-expected.txt

This is a testharness.js-based test.
[FAIL] ::part():dir() invalidation
  assert_equals: expected "rgb(0, 255, 0)" but got "rgb(255, 0, 0)"
[FAIL] ::part():dir() invalidation from setAttribute
  assert_equals: expected "rgb(0, 255, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
